add UnknownWordDetector interface
